update pVCU, regenerate code

+++ S1_VCU/TVS_26/Codegen/2025a_laptop/codegen/lib/vcu_step/minOrMax.h
@@ -9,7 +9,7 @@
 extern "C" {
 #endif
 
-void minimum(const double x[28], double ex[4]);
+void minimum(const float x[28], float ex[4]);
 
 #ifdef __cplusplus
 }
